Moves draftPoolApiClient into draftPage folder
